mobile_app/backend/requirements.txt
# Backend Server Requirements
# Python 3.9+

fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.0,<2.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
paho-mqtt>=1.6.0,<3.0.0
